src/common-lib/kafka/factory_test.go
// Package kafka implements kafka client configuration details
//
// Deprecated: kafka is old implementation of kafka connectivity and should not be used
// except for compatibility with legacy systems.
//
// Use https://gitlab.kksharmadevdev.com/platform/Platform-Infrastructure-lib/tree/master/messaging for all kafka connectivity
// This package is frozen and no new functionality will be added.
package kafka

import "testing"

func TestGetProducer(t *testing.T) {
	config := new(ProducerConfig)
	config.BrokerAddress = []string{"localhost:9092"}
	producerFactory := new(ProducerFactoryImpl)
	producer, _ := producerFactory.GetProducerService(*config)

	if producer == nil {
		t.Log("producer value is nil")
		t.Fail()
	}
}

func TestGetProducer_NilBrokerAddress(t *testing.T) {
	var config *ProducerConfig
	config = new(ProducerConfig)

	config.BrokerAddress = nil
	producerFactory := new(ProducerFactoryImpl)
	_, err := producerFactory.GetProducerService(*config)

	if err == nil || err.Error() != ErrorBrokerAddressNotProvided {
		t.Log("producer created with nil brokerAddress")
		t.Fail()
	}
}

func TestGetProducer_EmptyBrokerAddress(t *testing.T) {
	var config *ProducerConfig
	config = new(ProducerConfig)

	config.BrokerAddress = []string{}

	producerFactory := new(ProducerFactoryImpl)
	_, err := producerFactory.GetProducerService(*config)
	if err == nil || err.Error() != ErrorBrokerAddressNotProvided {
		t.Log("producer created with empty brokerAddress")
		t.Fail()
	}
}

func TestGetConsumer(t *testing.T) {

	config := new(ConsumerConfig)
	config.BrokerAddress = []string{"localhost:9092"}

	consumerFactory := new(ConsumerFactoryImpl)
	_, err := consumerFactory.GetConsumerService(*config)

	if err == nil || err.Error() != ErrorClientGroupIDNotProvided {
		t.Log("consumer value is nil")
		t.Fail()
	}
}

func TestGetConsumer_NilBrokerAddress(t *testing.T) {

	var config *ConsumerConfig
	config = new(ConsumerConfig)

	config.BrokerAddress = nil
	config.GroupID = "test101"
	consumerFactory := new(ConsumerFactoryImpl)
	_, err := consumerFactory.GetConsumerService(*config)
	if err == nil || err.Error() != ErrorBrokerAddressNotProvided {
		t.Log("Consumer created with nil brokerAddress")
		t.Fail()
	}
}
func TestGetConsumer_EmptyBrokerAddress(t *testing.T) {

	var config *ConsumerConfig
	config = new(ConsumerConfig)

	config.BrokerAddress = []string{}
	config.GroupID = "test101"
	consumerFactory := new(ConsumerFactoryImpl)
	_, err := consumerFactory.GetConsumerService(*config)

	if err == nil || err.Error() != ErrorBrokerAddressNotProvided {
		t.Log("Consumer created with empty brokerAddress")
		t.Fail()
	}
}

func TestGetConsumer_EmptyGroupID(t *testing.T) {

	var config *ConsumerConfig
	config = new(ConsumerConfig)

	config.BrokerAddress = []string{"localhost:9092"}
	config.GroupID = ""
	consumerFactory := new(ConsumerFactoryImpl)
	_, err := consumerFactory.GetConsumerService(*config)

	if err == nil || err.Error() != ErrorClientGroupIDNotProvided {
		t.Log("Consumer created with empty GroupID")
		t.Fail()
	}
}

func TestGetConsumer_EmptyTopics(t *testing.T) {

	var config *ConsumerConfig
	config = new(ConsumerConfig)

	config.BrokerAddress = []string{"localhost:9092"}
	config.GroupID = "grp1"

	consumerFactory := new(ConsumerFactoryImpl)
	_, err := consumerFactory.GetConsumerService(*config)

	if err == nil || err.Error() != ErrorTopicsNotProvided {
		t.Log("Consumer created with empty GroupID")
		t.Fail()
	}
}
func TestCommandFactory(t *testing.T) {
	cmdFactory := new(ConsumerCommandFactoryImpl)
	cmdService := cmdFactory.GetConsumerCommandService()
	if cmdService == nil {
		t.Fail()
	}
}
